delete the warning

+++ cdn/Init.cpp
@@ -60,7 +60,7 @@ void GraphLinkedTable::InsertEdgev_v(int v1, int v2, int b, int c)
 
 void GraphLinkedTable::PrintEdgeInfo()
 {
-	int i;
+	size_t i;
 	cout<<"下面打印边信息："<<endl;
 	for( i=0;i<vec_edge.size();i++)
 	{	cout<<vec_edge[i].in<<'\t'<<vec_edge[i].out<<'\t';
@@ -84,7 +84,7 @@ void GraphLinkedTable::PrintNetListInfo()
 	cout << "下面打印网络节点信息：" << endl;
 	for (int i = 0; i < v_count; i++)
 	{
-		for (int j = 0; j < vertex[i].edge_list.size(); j++)
+		for (size_t j = 0; j < vertex[i].edge_list.size(); j++)
 		{
 			cout << vertex[i].v_id << "\t" << vertex[i].edge_list[j].v_next << "\t";
 			cout << vertex[i].edge_list[j].bandwith << "\t" << vertex[i].edge_list[j].cost << endl;
@@ -122,7 +122,7 @@ stack<int> GraphLinkedTable::str_to_int(char *s)
 
 int GraphLinkedTable::GetEdgebw(int v1, int v2)
 {
-	int i = 0;
+	size_t i = 0;
 	while (  i<vertex[v1].edge_list.size()&&(vertex[v1].edge_list[i].v_next != v2))
 	{
 		i++;
@@ -136,7 +136,7 @@ int GraphLinkedTable::GetEdgebw(int v1, int v2)
 
 int GraphLinkedTable::GetEdgecost(int v1, int v2)
 {
-	int i = 0;
+	size_t i = 0;
 	while ( i<vertex[v1].edge_list.size()&& (vertex[v1].edge_list[i].v_next != v2))
 	{
 		i++;
@@ -151,7 +151,7 @@ int GraphLinkedTable::GetEdgecost(int v1, int v2)
 vector<int> GraphLinkedTable::GetNeiborNode(int v)
 {
 	vector<int> nextNode;
-	int i = 0;
+	size_t i = 0;
 	while (i < vertex[v].edge_list.size())
 	{
 		nextNode.push_back(vertex[v].edge_list[i].v_next);
@@ -207,23 +207,12 @@ void GraphLinkedTable::PrintHeader()
 		int v = vec_edge[e].out;
 		cout<<"e:"<<e<<" v:"<<v<<endl;
 	}
-	//cout<<"Now headEdge:"<<endl;
-	//for(int i=0;i<headEdge.size();i++)
-	//cout<<i<<'\t'<<headEdge[i]<<endl;
-	//cout<<"v2 Edege No:"<<endl;
-	//cout<<headEdge[2]<<endl;
-	//int formal=headEdge[2];
-	//while(vec_edge[formal].next!=-1)
-	//{
-	//cout<<vec_edge[formal].next<<endl;
-	//formal=vec_edge[formal].next;
-	//}
 }
 
 
 int GraphLinkedTable::ReturnEdgeNo(int in,int out)
 {
-	int i=0;
+	size_t i=0;
 	while(vec_edge.size()>i) 
 	{
 		if(vec_edge[i].in==in&& vec_edge[i].out==out)
